+++ docs/Whats-New.md
@@ -311,6 +311,7 @@ New:
 - OpenTopped transport rangefinding & deactivated state customizations (by Starkku)
 - Animation damage / weapon improvements (by Starkku)
 - Warhead self-damaging toggle (by Starkku)
+- Trailer animations inheriting owner (by Starkku)
 
 Vanilla fixes:
 - Fixed laser drawing code to allow for thicker lasers in house color draw mode (by Kerbiter, ChrisLv_CN)